Solution to Input and Print a Float

+++ Input and Print a Float@@6788d5012db25efec946f93a/code.c
@@ -1,8 +1,8 @@
 #include <stdio.h>
-float main(){
+int main(){
     float n;
     printf("");
     scanf("%f",&n);
     printf("You entered: %f/n",n);
-    return 0.00;
+    return 0;
 }
